Refactor account handling
